+++ src/circuit/gadget/ecc/chip/mul.rs
@@ -1,5 +1,5 @@
-use super::{add, CellValue, EccConfig, EccPoint, NonIdentityEccPoint, Var};
-use crate::{circuit::gadget::utilities::copy, constants::T_Q};
+use super::{add, CellValue, EccConfig, EccPoint, NonIdentityEccPoint, Var, T_Q};
+use crate::circuit::gadget::utilities::copy;
 use std::ops::{Deref, Range};
 
 use bigint::U256;
