+++ src/circuit/note_commit.rs
@@ -1444,8 +1444,8 @@ mod tests {
             },
         },
         constants::{
-            OrchardCommitDomains, OrchardFixedBases, OrchardHashDomains, L_ORCHARD_BASE, L_VALUE,
-            NOTE_COMMITMENT_PERSONALIZATION, T_Q,
+            fixed_bases::NOTE_COMMITMENT_PERSONALIZATION, OrchardCommitDomains, OrchardFixedBases,
+            OrchardHashDomains, L_ORCHARD_BASE, L_VALUE, T_Q,
         },
         primitives::sinsemilla::CommitDomain,
     };
